# openapiv2/laptop_service.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "laptop_service.proto",
    "version": "version not set"
  },
  "tags": [
    {
      "name": "LaptopService"
    }
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/v1/laptop/create": {
      "post": {
        "operationId": "LaptopService_CreateLaptop",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/protoCreateLaptopResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/protoCreateLaptopRequest"
            }
          }
        ],
        "tags": [
          "LaptopService"
        ]
      }
    },
    "/v1/laptop/rate": {
      "post": {
        "operationId": "LaptopService_RateLaptop",
        "responses": {
          "200": {
            "description": "A successful response.(streaming responses)",
            "schema": {
              "type": "object",
              "properties": {
                "result": {
                  "$ref": "#/definitions/protoRateLaptopResponse"
                },
                "error": {
                  "$ref": "#/definitions/rpcStatus"
                }
              },
              "title": "Stream result of protoRateLaptopResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "description": " (streaming inputs)",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/protoRateLaptopRequest"
            }
          }
        ],
        "tags": [
          "LaptopService"
        ]
      }
    },
    "/v1/laptop/search": {
      "get": {
        "operationId": "LaptopService_SearchLaptop",
        "responses": {
          "200": {
            "description": "A successful response.(streaming responses)",
            "schema": {
              "type": "object",
              "properties": {
                "result": {
                  "$ref": "#/definitions/protoSearchLaptopResponse"
                },
                "error": {
                  "$ref": "#/definitions/rpcStatus"
                }
              },
              "title": "Stream result of protoSearchLaptopResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "filter.maxPriceUsd",
            "in": "query",
            "required": false,
            "type": "number",
            "format": "double"
          },
          {
            "name": "filter.minCpuCores",
            "in": "query",
            "required": false,
            "type": "integer",
            "format": "int64"
          },
          {
            "name": "filter.minCpuGhz",
            "in": "query",
            "required": false,
            "type": "number",
            "format": "double"
          },
          {
            "name": "filter.minRam.value",
            "in": "query",
            "required": false,
            "type": "string",
            "format": "uint64"
          },
          {
            "name": "filter.minRam.unit",
            "in": "query",
            "required": false,
            "type": "string",
            "enum": [
              "UNKNOWN",
              "BIT",
              "BYTE",
              "KILOBYTE",
              "MEGABYTE",
              "GIGABYTE",
              "TERABYTE"
            ],
            "default": "UNKNOWN"
          }
        ],
        "tags": [
          "LaptopService"
        ]
      }
    },
    "/v1/laptop/upload_image": {
      "post": {
        "operationId": "LaptopService_UploadImage",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/protoUploadImageResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "description": " (streaming inputs)",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/protoUploadImageRequest"
            }
          }
        ],
        "tags": [
          "LaptopService"
        ]
      }
    }
  },
  "definitions": {
    "KeyboardLayout": {
      "type": "string",
      "enum": [
        "UNKNOWN",
        "QWERTY",
        "QWERTZ",
        "AZERTY"
      ],
      "default": "UNKNOWN"
    },
    "MemoryUnit": {
      "type": "string",
      "enum": [
        "UNKNOWN",
        "BIT",
        "BYTE",
        "KILOBYTE",
        "MEGABYTE",
        "GIGABYTE",
        "TERABYTE"
      ],
      "default": "UNKNOWN"
    },
    "ScreenPanel": {
      "type": "string",
      "enum": [
        "UNKNOWN",
        "IPS",
        "OLED"
      ],
      "default": "UNKNOWN"
    },
    "ScreenResolution": {
      "type": "object",
      "properties": {
        "width": {
          "type": "integer",
          "format": "int64"
        },
        "height": {
          "type": "integer",
          "format": "int64"
        }
      }
    },
    "StorageDriver": {
      "type": "string",
      "enum": [
        "UNKNOWN",
        "HDD",
        "SSD"
      ],
      "default": "UNKNOWN"
    },
    "protoCPU": {
      "type": "object",
      "properties": {
        "brand": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "numberCores": {
          "type": "integer",
          "format": "int64"
        },
        "numberThreads": {
          "type": "integer",
          "format": "int64"
        },
        "minGhz": {
          "type": "number",
          "format": "double"
        },
        "maxGhz": {
          "type": "number",
          "format": "double"
        }
      }
    },
    "protoCreateLaptopRequest": {
      "type": "object",
      "properties": {
        "laptop": {
          "$ref": "#/definitions/protoLaptop"
        }
      }
    },
    "protoCreateLaptopResponse": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        }
      }
    },
    "protoFilter": {
      "type": "object",
      "properties": {
        "maxPriceUsd": {
          "type": "number",
          "format": "double"
        },
        "minCpuCores": {
          "type": "integer",
          "format": "int64"
        },
        "minCpuGhz": {
          "type": "number",
          "format": "double"
        },
        "minRam": {
          "$ref": "#/definitions/protoMemory"
        }
      }
    },
    "protoGPU": {
      "type": "object",
      "properties": {
        "brand": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "minGhz": {
          "type": "number",
          "format": "double"
        },
        "maxGhz": {
          "type": "number",
          "format": "double"
        },
        "memory": {
          "$ref": "#/definitions/protoMemory"
        }
      }
    },
    "protoImageInfo": {
      "type": "object",
      "properties": {
        "laptopId": {
          "type": "string"
        },
        "imageType": {
          "type": "string"
        }
      }
    },
    "protoKeyboard": {
      "type": "object",
      "properties": {
        "layout": {
          "$ref": "#/definitions/KeyboardLayout"
        },
        "backlit": {
          "type": "boolean"
        }
      }
    },
    "protoLaptop": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "brand": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "cpu": {
          "$ref": "#/definitions/protoCPU"
        },
        "ram": {
          "$ref": "#/definitions/protoMemory"
        },
        "gpus": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/protoGPU"
          }
        },
        "storages": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/protoStorage"
          }
        },
        "screen": {
          "$ref": "#/definitions/protoScreen"
        },
        "keyboard": {
          "$ref": "#/definitions/protoKeyboard"
        },
        "weightKg": {
          "type": "number",
          "format": "double"
        },
        "weightLb": {
          "type": "number",
          "format": "double"
        },
        "priceUsd": {
          "type": "number",
          "format": "double"
        },
        "releaseYear": {
          "type": "integer",
          "format": "int64"
        },
        "updatedAt": {
          "type": "string",
          "format": "date-time"
        }
      }
    },
    "protoMemory": {
      "type": "object",
      "properties": {
        "value": {
          "type": "string",
          "format": "uint64"
        },
        "unit": {
          "$ref": "#/definitions/MemoryUnit"
        }
      }
    },
    "protoRateLaptopRequest": {
      "type": "object",
      "properties": {
        "laptopId": {
          "type": "string"
        },
        "score": {
          "type": "number",
          "format": "double"
        }
      }
    },
    "protoRateLaptopResponse": {
      "type": "object",
      "properties": {
        "laptopId": {
          "type": "string"
        },
        "ratedCount": {
          "type": "integer",
          "format": "int64"
        },
        "averageScore": {
          "type": "number",
          "format": "double"
        }
      }
    },
    "protoScreen": {
      "type": "object",
      "properties": {
        "sizeInch": {
          "type": "number",
          "format": "float"
        },
        "resolution": {
          "$ref": "#/definitions/ScreenResolution"
        },
        "panel": {
          "$ref": "#/definitions/ScreenPanel"
        },
        "multitouch": {
          "type": "boolean"
        }
      }
    },
    "protoSearchLaptopResponse": {
      "type": "object",
      "properties": {
        "laptop": {
          "$ref": "#/definitions/protoLaptop"
        }
      }
    },
    "protoStorage": {
      "type": "object",
      "properties": {
        "driver": {
          "$ref": "#/definitions/StorageDriver"
        },
        "memory": {
          "$ref": "#/definitions/protoMemory"
        }
      }
    },
    "protoUploadImageRequest": {
      "type": "object",
      "properties": {
        "info": {
          "$ref": "#/definitions/protoImageInfo"
        },
        "chunkData": {
          "type": "string",
          "format": "byte"
        }
      }
    },
    "protoUploadImageResponse": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "size": {
          "type": "integer",
          "format": "int64"
        }
      }
    },
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    }
  }
}